requirements.txt
aiogram == 2.15
python-dotenv == 0.19.1
asyncio == 3.4.3
docxtpl
openpyxl
asyncpg
